@@ -1,14 +1,3 @@
 package be.jago.ssehttp2.joke;
 
-public class ChuckNorrisJoke {
-
-    private String value;
-
-    public String getValue() {
-        return this.value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-}
+public record ChuckNorrisJoke(String value) {}
